+++ app/src/main/java/com/example/leafly_application_git/fragments/main_page/miljopoints/MiljopointsFragment.kt
@@ -153,7 +153,7 @@ class MiljopointsFragment : Fragment() {
                 textView_welcome_title.text = user?.username
                 progressbar_point_value.text = user?.balance.toString()
 
-                progressBar.setProgress(user!!.progress)
+                progressBar.setProgress(user!!.progress.toInt())
 
                 //Siple if/else to level up user when certain amount of progress is achieved
                 if (user?.progress!! >= 33){
